+++ Assets/Scripts/Spans/ForwardSpan/ForwardSpanImageDescription.cs
@@ -16,7 +16,7 @@ namespace Spans.ForwardSpan
             return GetRandomSprites();
         }
         
-        public override int GetRoundTime()
+        public override float GetRoundTime()
         {
             return currentRoundIndex * 3 + 2;
         }
